created a portfolio-layout for /about, moved the layout templates to /ui/layouts
