+++ imports/ui/layouts/portfolio-layout.html
@@ -1,0 +1,28 @@
+<template name="portfolioLayout">
+	<nav class="navbar navbar-default" role="navigation">
+		<div class="container-fluid">
+			<div class="navbar-header">
+				<button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#navbar" aria-expanded="false" aria-controls="navbar">
+					<span class="sr-only">Toggle navigation</span>
+					<span class="icon-bar"></span>
+					<span class="icon-bar"></span>
+					<span class="icon-bar"></span>
+				</button>
+				<a class="navbar-brand">The Kove</a>
+			</div>
+			<div id="navbar" class="navbar-collapse collapse">
+				<ul class="nav navbar-nav"></ul>
+				<ul class="nav navbar-nav navbar-right">
+					<li><a href="/blog" class="nav-link">Blog</a></li>
+					<li><a href="/about" class="nav-link">About</a></li>
+				</ul>
+			</div>
+		</div>
+	</nav>
+	<div class="container-fluid">
+
+		{{> Template.dynamic template=content}}
+
+	</div>
+	<footer>footer goes here</footer>
+</template>
